 requirements.txt
Flask==3.0
supabase==2.0
python-dotenv==1.0